requirements/base.txt
# Python Utilities
python-slugify==8.0.4
Pillow==11.1.0
argon2-cffi==23.1.0
pydantic==2.10.6
pydantic-settings==2.7.1
boto3==1.36.24

# Redis
redis==5.2.1
hiredis==3.1.0
psycopg[pool]==3.2.6

# Django and Extensions
django==5.1.6
django-model-utils==5.0.0
django-redis==5.4.0
django-import-export==4.3.5
openpyxl==3.1.5
django-phonenumber-field[phonenumbers]==8.0.0
django-polymorphic==3.1.0
django-unfold==0.49.1
fcm-django==2.2.1
django-solo==2.4.0
rules==3.5
django-imagekit==5.0.0
django-filter==24.3
django-money==3.5.3
django-cors-headers==4.7.0
dj-database-url==2.3.0
django-model-suite==0.1.15

# Django REST Framework and Extensions
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
drf-spectacular==0.28.0
drf-standardized-errors[openapi]==0.14.1
